refactor: RR4 route hocs
