+++ routes-v1/hocs/index.js
@@ -1,0 +1,3 @@
+export { default as withBackground } from './with-background'
+export { default as withSidebar } from './with-sidebar'
+export { default as withUser } from './with-user'
