+++ routes-v1/hocs/with-background.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import BackgroundContainer from '~routes/admin/not-authenticated/background.connected'
+
+const withBackground = Component => props => (
+  <BackgroundContainer {...props}>
+    <Component {...props} />
+  </BackgroundContainer>
+)
+
+export default withBackground
