+++ routes-v1/hocs/with-user.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import CurrentUserContainer from '~routes/admin/authenticated/container'
+
+const withUser = Component => props => (
+  <CurrentUserContainer {...props}>
+    <Component {...props} />
+  </CurrentUserContainer>
+)
+
+export default withUser
