+++ routes-v1/index.js
@@ -7,7 +7,6 @@ import {
   withRouter,
   Switch
 } from 'react-router-dom'
-import { connect } from 'react-redux'
 
 import '~client/styles/main.scss'
 
@@ -20,62 +19,7 @@ import AccountLogin from '~routes/admin/not-authenticated/account-login/page.con
 import CommunityList from '~routes/admin/authenticated/external/community-list/page.connected'
 import MobilizationsList from '~routes/admin/authenticated/sidebar/mobilizations-list/page.connected'
 
-const withBackground = Component => props => (
-  <BackgroundContainer {...props}>
-    <Component {...props} />
-  </BackgroundContainer>
-)
-
-const withUser = Component => props => (
-  <CurrentUserContainer {...props}>
-    <Component {...props} />
-  </CurrentUserContainer>
-)
-
-
-import MobSelectors from '~client/mobrender/redux/selectors'
-import * as MobActions from '~client/mobrender/redux/action-creators'
-import * as CommunitySelectors from '~client/community/selectors'
-import DNSControlSelectors from '~client/community/dns-control-selectors'
-import * as DNSControlActions from '~client/community/action-creators/dns-control'
-import SidebarModule from '~client/components/navigation/sidebar'
-import { Loading } from '~client/components/await'
-
-const mapStateToSidebarProps = (state, props) => ({
-  community: CommunitySelectors.getCurrent(state),
-  loading: MobSelectors(state, props).mobilizationsIsLoading(),
-  loaded: MobSelectors(state, props).mobilizationsIsLoaded(),
-  sidebarProps: SidebarModule.getSidebarProps(state, props),
-  dnsControlSelectors: DNSControlSelectors(state)
-})
-const mapDispatchToSidebarProps = { ...MobActions, ...DNSControlActions }
-const withSidebar = Component => connect(mapStateToSidebarProps, mapDispatchToSidebarProps)(
-  class SidebarHOC extends React.Component {
-    componentDidMount () {
-      const promises = []
-      const community = this.props.community
-      const shouldMakeFetch = community && !this.props.loaded && !this.props.loading
-
-      shouldMakeFetch && promises.push(this.props.asyncFetchMobilizations(community.id))
-      !this.props.dnsControlSelectors.dnsHostedZones().isLoaded() && promises.push(
-        this.props.asyncFetchHostedZones()
-      )
-      return Promise.all(promises)
-    }
-
-    render () {
-      return (
-        <div>
-          {this.props.loading ? <Loading /> : (
-            <SidebarModule.Sidebar {...this.props.sidebarProps}>
-              <Component {...this.props} />
-            </SidebarModule.Sidebar>
-          )}
-        </div>
-      )
-    }
-  }
-)
+import { withBackground, withUser, withSidebar } from '~root/routes-v1/hocs'
 
 const About = () => (
   <div>
